ADD auto blocked user redirect
Requested not to direct blocked users to the welcome page, instead
sending them to a different page that'll let them know where to get help
if the block was in error. The welcome page redirect is only for newly
created users. But the blocked user redirect will be for every blocked
user login, not just on create.

This expects $wgAutoBlockedUserRedirect to be configured in
LocalSettings.php with the name of the page for blocked users.